Fixed JMotionDevice

+++ source/cpp/toadlet/flick/plugins/jmotiondevice/us_toadlet_flick_NMotionDeviceListener.cpp
@@ -9,7 +9,7 @@ using namespace toadlet::flick;
 
 JNIEXPORT void JNICALL Java_us_toadlet_flick_NMotionDeviceListener_motionDetected(JNIEnv *env,jobject obj,jobject dataObj){
 	MotionData data;
-	data.time=env->GetIntField(dataObj,timeID);
+	data.time=env->GetLongField(dataObj,timeID);
 	env->GetFloatArrayRegion((jfloatArray)env->GetObjectField(dataObj,accelerationID),0,3,data.acceleration.getData());
 	env->GetFloatArrayRegion((jfloatArray)env->GetObjectField(dataObj,velocityID),0,3,data.velocity.getData());
